refactor: separate List extension functions to extensions.kt
